+++ src/lib/mat-multi-sort.directive.spec.ts
@@ -1,7 +1,15 @@
 import { Component } from "@angular/core";
-import { ComponentFixture, TestBed } from "@angular/core/testing";
-import { MatSortable } from "@angular/material/sort";
-import { MatMultiSortDirective } from "./mat-multi-sort.directive";
+import {
+  ComponentFixture,
+  fakeAsync,
+  TestBed,
+  tick,
+} from "@angular/core/testing";
+import { MatSortable, Sort } from "@angular/material/sort";
+import {
+  MatMultiSortDirective,
+  SORT_PERSISTENCE_STORAGE,
+} from "./mat-multi-sort.directive";
 
 @Component({
   selector: "mat-multi-sort-test",
@@ -10,13 +18,17 @@ import { MatMultiSortDirective } from "./mat-multi-sort.directive";
 class TestComponent extends MatMultiSortDirective {}
 
 describe("MatMultiSortDirective", () => {
+  let setItemSpy: jasmine.Spy;
   let fixture: ComponentFixture<TestComponent>;
-  let directive: MatMultiSortDirective;
+  let directive: TestComponent;
 
   beforeEach(async () => {
+    spyOn(globalThis.sessionStorage, "getItem");
+    setItemSpy = spyOn(globalThis.sessionStorage, "setItem");
     await TestBed.configureTestingModule({
       imports: [TestComponent, MatMultiSortDirective],
-      providers: [],
+        { provide: SORT_PERSISTENCE_STORAGE, useValue: sessionStorage },
+      ],
     }).compileComponents();
 
     fixture = TestBed.createComponent(TestComponent);
@@ -61,6 +73,7 @@ describe("MatMultiSortDirective", () => {
     expect(directive.active).toBe("test");
     expect(directive.direction).toBe("desc");
     expect(spy).toHaveBeenCalledWith({ active: "test", direction: "desc" });
+    expect(setItemSpy).toHaveBeenCalled();
   });
 
   it("should append a new column to the list of sort columns with the default sort direction unless otherwise specified", () => {
@@ -75,6 +88,7 @@ describe("MatMultiSortDirective", () => {
     expect(directive.active).toBe("test");
     expect(directive.direction).toBe("asc");
     expect(spy).toHaveBeenCalledWith({ active: "test", direction: "asc" });
+    expect(setItemSpy).toHaveBeenCalled();
   });
 
   it("should update the direction of an existing column", () => {
@@ -88,6 +102,7 @@ describe("MatMultiSortDirective", () => {
     expect(directive.active).toBe("col1");
     expect(directive.direction).toBe("desc");
     expect(spy).toHaveBeenCalledWith({ active: "col1", direction: "desc" });
+    expect(setItemSpy).toHaveBeenCalled();
   });
 
   it("should remove the direction of an existing column if the direction is empty", () => {
@@ -101,6 +116,7 @@ describe("MatMultiSortDirective", () => {
     expect(directive.active).toBe("col2");
     expect(directive.direction).toBe("");
     expect(spy).toHaveBeenCalledWith({ active: "col2", direction: "" });
+    expect(setItemSpy).toHaveBeenCalled();
   });
 
   it("should remove a sort level by its identifier", () => {
@@ -110,7 +126,8 @@ describe("MatMultiSortDirective", () => {
       { active: "col1", direction: "asc" },
       { active: "col3", direction: "asc" },
     ]);
-    expect(spy).toHaveBeenCalledWith();
+    expect(spy).toHaveBeenCalled();
+    expect(setItemSpy).toHaveBeenCalled();
   });
 
   it("should not change the sort levels if the identifier is not found", () => {
@@ -122,6 +139,7 @@ describe("MatMultiSortDirective", () => {
       { active: "col3", direction: "asc" },
     ]);
     expect(spy).not.toHaveBeenCalled();
+    expect(setItemSpy).not.toHaveBeenCalled();
   });
 
   it("should reorder the sort levels when previousIndex and currentIndex are different", () => {
@@ -133,6 +151,7 @@ describe("MatMultiSortDirective", () => {
       { active: "col1", direction: "asc" },
     ]);
     expect(spy).toHaveBeenCalledWith({ active: "col1", direction: "asc" });
+    expect(setItemSpy).toHaveBeenCalled();
   });
 
   it("should not reorder the sort levels when previousIndex and currentIndex are the same", () => {
@@ -144,6 +163,7 @@ describe("MatMultiSortDirective", () => {
       { active: "col3", direction: "asc" },
     ]);
     expect(spy).not.toHaveBeenCalled();
+    expect(setItemSpy).not.toHaveBeenCalled();
   });
 
   it("should toggle the sort direction for an existing column", () => {
@@ -157,6 +177,7 @@ describe("MatMultiSortDirective", () => {
     expect(directive.active).toBe("col1");
     expect(directive.direction).toBe("desc");
     expect(spy).toHaveBeenCalledWith({ active: "col1", direction: "desc" });
+    expect(setItemSpy).toHaveBeenCalled();
   });
 
   it("should not change the sort direction for a column that is not sorted", () => {
@@ -168,6 +189,7 @@ describe("MatMultiSortDirective", () => {
       { active: "col3", direction: "asc" },
     ]);
     expect(spy).not.toHaveBeenCalled();
+    expect(setItemSpy).not.toHaveBeenCalled();
   });
 
   it("should cycle through sort directions for an existing column", () => {
@@ -183,6 +205,7 @@ describe("MatMultiSortDirective", () => {
     expect(directive.active).toBe("col1");
     expect(directive.direction).toBe("desc");
     expect(spy).toHaveBeenCalledWith({ active: "col1", direction: "desc" });
+    expect(setItemSpy).toHaveBeenCalledTimes(1);
 
     directive.toggleSortDirection("col1");
     expect(directive._sorts()).toEqual([
@@ -193,6 +216,7 @@ describe("MatMultiSortDirective", () => {
     expect(directive.active).toBe("col1");
     expect(directive.direction).toBe("asc");
     expect(spy).toHaveBeenCalledWith({ active: "col1", direction: "asc" });
+    expect(setItemSpy).toHaveBeenCalledTimes(2);
   });
 
   it("should clear the current sorting state", () => {
@@ -202,5 +226,37 @@ describe("MatMultiSortDirective", () => {
     expect(directive.direction).toBe("");
     expect(directive._sorts()).toEqual([]);
     expect(spy).toHaveBeenCalled();
+    expect(setItemSpy).toHaveBeenCalled();
+  });
+
+  it("should trigger effect", fakeAsync(() => {
+    const spy = spyOn(directive.sortChange, "emit");
+    directive._sorts.update((e) => e.reverse());
+    fixture.detectChanges();
+    tick();
+    expect(spy).toHaveBeenCalledWith({ active: "col1", direction: "asc" });
+  }));
+});
+
+describe("MatMultiSortDirective", () => {
+  it("should restore the persisted settings", async () => {
+    const test: Sort[] = [
+      { active: "col1", direction: "asc" },
+      { active: "col2", direction: "desc" },
+      { active: "col3", direction: "asc" },
+    ];
+    spyOn(globalThis.sessionStorage, "getItem").and.returnValue(
+      JSON.stringify(test)
+    );
+    await TestBed.configureTestingModule({
+      imports: [TestComponent, MatMultiSortDirective],
+      providers: [
+        { provide: SORT_PERSISTENCE_STORAGE, useValue: sessionStorage },
+      ],
+    }).compileComponents();
+
+    const fixture = TestBed.createComponent(TestComponent);
+    const directive = fixture.componentInstance;
+    expect(directive._sorts()).toEqual(test);
   });
 });