+++ src/lib/mat-table-column-config-trigger.directive.ts
@@ -5,11 +5,8 @@ import {
   Directive,
   ElementRef,
   HostListener,
-  Injector,
-  Input,
   ViewContainerRef,
 } from "@angular/core";
-import { TABLE_COLUMNS, TableColumn } from "./mat-table-column-config";
 import { MatTableColumnConfigComponent } from "./mat-table-column-config/mat-table-column-config.component";
 
 @Directive({
@@ -20,15 +17,6 @@ import { MatTableColumnConfigComponent } from "./mat-table-column-config/mat-tab
 export class MatTableColumnConfigTriggerDirective<T> {
   private _componentRef: ComponentRef<MatTableColumnConfigComponent<T>> | null =
     null;
-  /**
-   * Input property that accepts an array of table column configurations.
-   * The alias for this input property is "matTableColumnConfigTrigger".
-   * This property is required.
-   *
-   * @type {TableColumn<T>[]} columns - The array of table column configurations.
-   */
-  @Input({ alias: "matTableColumnConfigTrigger", required: true })
-  columns!: TableColumn<T>[];
 
   /**
    * Gets the reference to the MatTableColumnConfigComponent.
@@ -45,6 +33,7 @@ export class MatTableColumnConfigTriggerDirective<T> {
     private readonly overlay: Overlay,
     private readonly viewContainerRef: ViewContainerRef
   ) {}
+
   @HostListener("click")
   onClick(): void {
     // Create the component portal
@@ -68,14 +57,10 @@ export class MatTableColumnConfigTriggerDirective<T> {
       hasBackdrop: true,
       backdropClass: "cdk-overlay-transparent-backdrop",
     });
-    const injector = Injector.create({
-      providers: [{ provide: TABLE_COLUMNS, useValue: this.columns }],
-      parent: this.viewContainerRef.injector,
-    });
     const portal = new ComponentPortal(
       MatTableColumnConfigComponent<T>,
       this.viewContainerRef,
-      injector
+      this.viewContainerRef.injector
     );
     this._componentRef = overlayRef.attach(portal);