+++ src/lib/mat-table-column-config.ts
@@ -12,13 +12,31 @@ export interface TableColumn<T> {
 }
 
 /**
- * Injection token for providing table column configurations.
+ * Injection token for the storage mechanism used to persist column configuration.
  *
- * This token is used to inject an array of `TableColumn` configurations
- * into Angular components or services. The generic type `unknown` is used
- * to allow for flexibility in the type of data that can be represented
- * by the table columns.
+ * This token can be used to provide a custom storage implementation for saving
+ * and retrieving the state of table column configurations.
+ *
  */
-export const TABLE_COLUMNS = new InjectionToken<TableColumn<unknown>>(
-  "TABLE_COLUMNS"
+export const COLUMN_CONFIG_PERSISTENCE_STORAGE = new InjectionToken<Storage>(
+  "COLUMN_CONFIG_PERSISTENCE_STORAGE"
+);
+
+/**
+ * Injection token used to enable or disable column configuration persistence.
+ *
+ * This token can be provided with a boolean value to indicate whether the
+ * column configurations should be persisted (e.g., in local storage or a database).
+ *
+ */
+export const COLUMN_CONFIG_PERSISTENCE_ENABLED = new InjectionToken<boolean>(
+  "COLUMN_CONFIG_PERSISTENCE_ENABLED"
+);
+
+/**
+ * Injection token for the column configuration persistence key.
+ * This token is used to provide a unique key for persisting column configurations.
+ */
+export const COLUMN_CONFIG_PERSISTENCE_KEY = new InjectionToken<string>(
+  "COLUMN_CONFIG_PERSISTENCE_KEY"
 );