+++ src/lib/mat-table-column-config/mat-table-column-config.component.spec.ts
@@ -1,19 +1,26 @@
 import { CdkDragDrop } from "@angular/cdk/drag-drop";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { generateColumns, Test } from "../../test";
-import { TABLE_COLUMNS } from "../mat-table-column-config";
+import { COLUMN_CONFIG_PERSISTENCE_ENABLED } from "../mat-table-column-config";
+import { MatTableColumnConfigPersistenceService } from "../mat-table-column-config-persistence.service";
 import { MatTableColumnConfigComponent } from "./mat-table-column-config.component";
 
 describe("MatTableColumnConfigComponent", () => {
   let component: MatTableColumnConfigComponent<Test>;
   let fixture: ComponentFixture<MatTableColumnConfigComponent<Test>>;
+  let service: MatTableColumnConfigPersistenceService<Test>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [MatTableColumnConfigComponent],
-      providers: [{ provide: TABLE_COLUMNS, useValue: generateColumns() }],
+      providers: [
+        MatTableColumnConfigPersistenceService,
+        { provide: COLUMN_CONFIG_PERSISTENCE_ENABLED, useValue: false },
+      ],
     }).compileComponents();
 
+    service = TestBed.inject(MatTableColumnConfigPersistenceService);
+    service.columns = generateColumns();
     fixture = TestBed.createComponent(MatTableColumnConfigComponent<Test>);
     component = fixture.componentInstance;
     fixture.detectChanges();
@@ -29,11 +36,15 @@ describe("MatTableColumnConfigComponent", () => {
       currentIndex: 1,
     } as CdkDragDrop<Test>;
 
-    let order = component.columns.map((e) => e.id);
-    expect(order).toEqual(["id", "name", "value"]);
+    let orderComponent = component.columns.map((e) => e.id);
+    let orderService = service.columns.map((e) => e.id);
+    expect(orderComponent).toEqual(["id", "name", "value"]);
+    expect(orderService).toEqual(orderComponent);
     component.onColumnDropped(event);
-    order = component.columns.map((e) => e.id);
-    expect(order).toEqual(["name", "id", "value"]);
+    orderComponent = component.columns.map((e) => e.id);
+    orderService = service.columns.map((e) => e.id);
+    expect(orderComponent).toEqual(["name", "id", "value"]);
+    expect(orderService).toEqual(orderComponent);
   });
 
   it("should toggle column visibility", () => {