+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- [#36](https://github.com/pgerke/ngx-mat-table-multi-sort/issues/36): Sort and column configuration persistence
+- [#40](https://github.com/pgerke/ngx-mat-table-multi-sort/issues/40): Remove `!important` from SCSS rule in `mat-multi-sort-control` component
+
 ## [18.1.0](https://github.com/pgerke/ngx-mat-table-multi-sort/releases/tag/18.1.0) - 22.01.2025
 
 ### Added